requirements.txt
qgis-stubs
ipykernel